move 'mk' functions from LanguageRenderer to new Constructors file
