+++ code/drasil-gool/drasil-gool.cabal
@@ -15,6 +15,7 @@ library
     , GOOL.Drasil.Helpers
     , GOOL.Drasil.CodeAnalysis
     , GOOL.Drasil.CodeInfo
+    , GOOL.Drasil.LanguageRenderer.Constructors
     , GOOL.Drasil.LanguageRenderer.LanguagePolymorphic
     , GOOL.Drasil.LanguageRenderer
     , GOOL.Drasil.LanguageRenderer.JavaRenderer
@@ -56,6 +57,7 @@ executable codegenTest
     , GOOL.Drasil.Helpers
     , GOOL.Drasil.CodeAnalysis
     , GOOL.Drasil.CodeInfo
+    , GOOL.Drasil.LanguageRenderer.Constructors
     , GOOL.Drasil.LanguageRenderer.LanguagePolymorphic
     , GOOL.Drasil.LanguageRenderer
     , GOOL.Drasil.LanguageRenderer.JavaRenderer
